requirements.txt
Django==1.6.3
Markdown==2.4
MySQL-python
South==0.8.4
beautifulsoup4==4.3.2
boto==2.24.0
django-appconf==0.6
django-bower==4.8.1
django-compressor==1.3
django-filter==0.7
django-post-office==0.8.3
django-tastypie==0.11.0
djangorestframework==2.3.13
httplib2==0.8
jsonfield==0.9.20
paypalrestsdk==0.6.4
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
pycrypto==2.6.1
reportlab
pytz
django-mailchimp-v1.3
pisa
html5lib
pypdf
